Extract shared functionality into separate module
This is necessary as the functionality will be used in the update
helper.

Also contains a few minor improvements for the CMake config, along with
the necessary changes.
