+++ src/CMakeLists.txt
@@ -1,4 +1,5 @@
-cmake_minimum_required(VERSION 3.2)
+# in version 3.6 IMPORTED_TARGET has been added to pkg_*_modules
+cmake_minimum_required(VERSION 3.6)
 
 set(CMAKE_CXX_STANDARD 11)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -6,14 +7,20 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 find_package(Qt5 REQUIRED COMPONENTS Core Widgets)
 
 find_package(PkgConfig)
-pkg_check_modules(GLIB REQUIRED glib-2.0>=2.40)
-include_directories(${GLIB_INCLUDE_DIRS})
+pkg_check_modules(glib REQUIRED glib-2.0>=2.40 IMPORTED_TARGET)
 
-add_executable(AppImageLauncher main.cpp)
-target_link_libraries(AppImageLauncher Qt5::Core Qt5::Widgets ${GLIB_LIBRARIES} libappimage xdg-basedir)
-target_compile_definitions(AppImageLauncher
+
+# utility library
+add_library(shared STATIC shared.h shared.cpp)
+target_link_libraries(shared PUBLIC PkgConfig::glib Qt5::Core Qt5::Widgets xdg-basedir libappimage)
+target_compile_definitions(shared
     PRIVATE -DCMAKE_INSTALL_PREFIX="${CMAKE_INSTALL_PREFIX}"
 )
+
+
+# main AppImageLauncher application
+add_executable(AppImageLauncher main.cpp)
+target_link_libraries(AppImageLauncher shared Qt5::Core Qt5::Widgets PkgConfig::glib libappimage)
 # set binary runtime rpath to make sure the libappimage.so built and installed by this project is going to be used
 # by the installed binaries (be it the .deb, the AppImage, or whatever)
 # in order to make the whole install tree relocatable, a relative path is used
@@ -30,6 +37,7 @@ install(
 )
 
 
+# AppImage removal helper
 add_executable(remove remove_main.cpp)
 target_link_libraries(remove Qt5::Core Qt5::Widgets libappimage)
 # see AppImageLauncher for a description
@@ -41,6 +49,7 @@ install(
 )
 
 
+# AppImage update helper
 add_executable(update update_main.cpp)
 target_link_libraries(update Qt5::Core Qt5::Widgets libappimage libappimageupdate-qt)
 # see AppImageLauncher for a description
